Integer ID -> Long ID

+++ dao-orm/src/main/java/com/training/carsharing/repository/BrandRepository.java
@@ -3,6 +3,6 @@ package com.training.carsharing.repository;
 import com.training.carsharing.model.impl.Brand;
 
 //@Repository
-public interface BrandRepository extends AbstractRepository<Brand, Integer> {
+public interface BrandRepository extends AbstractRepository<Brand, Long> {
 
 }
