+++ dao-orm/src/main/java/com/training/carsharing/repository/CarParameterRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.CarParameter;
 
-public interface CarParameterRepository extends AbstractRepository<CarParameter, Integer> {
+public interface CarParameterRepository extends AbstractRepository<CarParameter, Long> {
 
 }
