+++ dao-orm/src/main/java/com/training/carsharing/repository/CarsPhotoRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.CarsPhoto;
 
-public interface CarsPhotoRepository extends AbstractRepository<CarsPhoto, Integer> {
+public interface CarsPhotoRepository extends AbstractRepository<CarsPhoto, Long> {
 }
