+++ dao-orm/src/main/java/com/training/carsharing/repository/FuelRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Fuel;
 
-public interface FuelRepository extends AbstractRepository<Fuel, Integer> {
+public interface FuelRepository extends AbstractRepository<Fuel, Long> {
 
 }
