+++ dao-orm/src/main/java/com/training/carsharing/repository/GearboxRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Gearbox;
 
-public interface GearboxRepository extends AbstractRepository<Gearbox, Integer> {
+public interface GearboxRepository extends AbstractRepository<Gearbox, Long> {
 
 }
