+++ dao-orm/src/main/java/com/training/carsharing/repository/UserAccountRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.UserAccount;
 
-public interface UserAccountRepository extends AbstractRepository<UserAccount, Integer> {
+public interface UserAccountRepository extends AbstractRepository<UserAccount, Long> {
 }
