+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/AdRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Ad;
 
-public interface AdRepositoryCustom extends AbstractRepositoryCustom<Ad, Integer> {
+public interface AdRepositoryCustom extends AbstractRepositoryCustom<Ad, Long> {
 }
