+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/BodyTypeRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.BodyType;
 
-public interface BodyTypeRepositoryCustom extends AbstractRepositoryCustom<BodyType, Integer> {
+public interface BodyTypeRepositoryCustom extends AbstractRepositoryCustom<BodyType, Long> {
 
 }
