+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/CarParameterRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.CarParameter;
 
-public interface CarParameterRepositoryCustom extends AbstractRepositoryCustom<CarParameter, Integer> {
+public interface CarParameterRepositoryCustom extends AbstractRepositoryCustom<CarParameter, Long> {
 
 }
