+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/CarRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Car;
 
-public interface CarRepositoryCustom extends AbstractRepositoryCustom<Car, Integer> {
+public interface CarRepositoryCustom extends AbstractRepositoryCustom<Car, Long> {
 }
