+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/CarsPhotoRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.CarsPhoto;
 
-public interface CarsPhotoRepositoryCustom extends AbstractRepositoryCustom<CarsPhoto, Integer> {
+public interface CarsPhotoRepositoryCustom extends AbstractRepositoryCustom<CarsPhoto, Long> {
 }
