+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/DrivingLicenseRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.DrivingLicense;
 
-public interface DrivingLicenseRepositoryCustom extends AbstractRepositoryCustom<DrivingLicense, Integer> {
+public interface DrivingLicenseRepositoryCustom extends AbstractRepositoryCustom<DrivingLicense, Long> {
 
 }
