+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/EngineTypeRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.EngineType;
 
-public interface EngineTypeRepositoryCustom extends AbstractRepositoryCustom<EngineType, Integer> {
+public interface EngineTypeRepositoryCustom extends AbstractRepositoryCustom<EngineType, Long> {
 
 }
