+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/GearboxRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Gearbox;
 
-public interface GearboxRepositoryCustom extends AbstractRepositoryCustom<Gearbox, Integer> {
+public interface GearboxRepositoryCustom extends AbstractRepositoryCustom<Gearbox, Long> {
 
 }
