+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/PassportRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Passport;
 
-public interface PassportRepositoryCustom extends AbstractRepositoryCustom<Passport, Integer> {
+public interface PassportRepositoryCustom extends AbstractRepositoryCustom<Passport, Long> {
 
 }
