+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/UserAccountRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.UserAccount;
 
-public interface UserAccountRepositoryCustom extends AbstractRepositoryCustom<UserAccount, Integer> {
+public interface UserAccountRepositoryCustom extends AbstractRepositoryCustom<UserAccount, Long> {
 }
