+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/CarParameterRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class CarParameterRepositoryImpl extends AbstractRepositoryImpl<CarParameter, Integer> implements CarParameterRepositoryCustom {
+public class CarParameterRepositoryImpl extends AbstractRepositoryImpl<CarParameter, Long> implements CarParameterRepositoryCustom {
 
     protected CarParameterRepositoryImpl() { super(CarParameter.class); }
 
@@ -27,7 +27,7 @@ public class CarParameterRepositoryImpl extends AbstractRepositoryImpl<CarParame
 
     //TODO delete methods
     @Override
-    public CarParameter findOneFullInfo(final Integer id) {
+    public CarParameter findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<CarParameter> cq = cb.createQuery(CarParameter.class);
