+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/DriveRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class DriveRepositoryImpl extends AbstractRepositoryImpl<Drive, Integer> implements DriveRepositoryCustom {
+public class DriveRepositoryImpl extends AbstractRepositoryImpl<Drive, Long> implements DriveRepositoryCustom {
 
     protected DriveRepositoryImpl() {
         super(Drive.class);
@@ -29,7 +29,7 @@ public class DriveRepositoryImpl extends AbstractRepositoryImpl<Drive, Integer> 
 
     //TODO delete methods
     @Override
-    public Drive findOneFullInfo(final Integer id) {
+    public Drive findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<Drive> cq = cb.createQuery(Drive.class);
