+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/EngineTypeRepositoryImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
 
 
 @Repository
-public class EngineTypeRepositoryImpl extends AbstractRepositoryImpl<EngineType, Integer> implements EngineTypeRepositoryCustom {
+public class EngineTypeRepositoryImpl extends AbstractRepositoryImpl<EngineType, Long> implements EngineTypeRepositoryCustom {
 
     protected EngineTypeRepositoryImpl() {
         super(EngineType.class);
@@ -29,7 +29,7 @@ public class EngineTypeRepositoryImpl extends AbstractRepositoryImpl<EngineType,
     }
 
     @Override
-    public EngineType findOneFullInfo(Integer id) {
+    public EngineType findOneFullInfo(Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<EngineType> cq = cb.createQuery(EngineType.class);
