+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/FuelRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class FuelRepositoryImpl extends AbstractRepositoryImpl<Fuel, Integer> implements FuelRepositoryCustom {
+public class FuelRepositoryImpl extends AbstractRepositoryImpl<Fuel, Long> implements FuelRepositoryCustom {
 
     protected FuelRepositoryImpl() {
         super(Fuel.class);
@@ -29,7 +29,7 @@ public class FuelRepositoryImpl extends AbstractRepositoryImpl<Fuel, Integer> im
 
     //TODO delete methods
     @Override
-    public Fuel findOneFullInfo(final Integer id) {
+    public Fuel findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<Fuel> cq = cb.createQuery(Fuel.class);
